Add simplistic "falling sand" behavior for particles
Also:
- Add debug canvas for messing with some particle settings
- Move particle logic into the particle class
- Add the ability to "paint" with sand and concrete using the middle mouse button and fourth mouse button.
- And just some general messing about.

Note: The reason the falling sand behavior is simplistic is because I'm only applying it to falling/flying particles, and as soon as a particle collides and comes to a rest it's taken out of the simulation. This is very noticeable when painting with sand as most of the sand will just disappear and in general doesn't behave like you'd expect.

Ref. #1